Make the script skip the main module
There is no point in processing a module for which we are updating
dependencies, as we won't be able to update its version anyway.

Yes, this module will still be filtered in the second stage, but there
is no point in including it in the list of modules to search for
updates.